requirements.txt
Django
djangorestframework==3.14.0
uritemplate==4.1.1
pyyaml
